fix call access event
